feat(jdbc): Rename JdbcAuthentication, add JdbcRequestData
