+++ connectors/jdbc/src/main/java/io/camunda/connector/jdbc/model/request/JdbcRequest.java
@@ -1,0 +1,31 @@
+/*
+ * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH
+ * under one or more contributor license agreements. Licensed under a proprietary license.
+ * See the License.txt file for more information. You may not use this file
+ * except in compliance with the proprietary license.
+ */
+package io.camunda.connector.jdbc.model.request;
+
+import static io.camunda.connector.generator.java.annotation.TemplateProperty.PropertyType.Dropdown;
+
+import io.camunda.connector.generator.java.annotation.TemplateProperty;
+import io.camunda.connector.jdbc.model.request.auth.JdbcConnection;
+import jakarta.validation.Valid;
+import jakarta.validation.constraints.NotNull;
+
+public record JdbcRequest(
+    @NotNull
+        @TemplateProperty(
+            id = "database",
+            label = "Database",
+            group = "database",
+            type = Dropdown,
+            constraints = @TemplateProperty.PropertyConstraints(notEmpty = true),
+            choices = {
+              @TemplateProperty.DropdownPropertyChoice(label = "MSSQL", value = "MSSQL"),
+              @TemplateProperty.DropdownPropertyChoice(label = "MySQL", value = "MYSQL"),
+              @TemplateProperty.DropdownPropertyChoice(label = "PostgreSQL", value = "POSTGRESQL"),
+            })
+        SupportedDatabase database,
+    @Valid JdbcConnection authentication,
+    @Valid @NotNull JdbcRequestData data) {}
