+++ connectors/jdbc/src/main/java/io/camunda/connector/jdbc/model/request/JdbcRequestData.java
@@ -4,34 +4,16 @@
  * See the License.txt file for more information. You may not use this file
  * except in compliance with the proprietary license.
  */
-package io.camunda.connector.jdbc.model;
-
-import static io.camunda.connector.generator.java.annotation.TemplateProperty.PropertyType.Dropdown;
+package io.camunda.connector.jdbc.model.request;
 
 import io.camunda.connector.feel.annotation.FEEL;
 import io.camunda.connector.generator.dsl.Property;
 import io.camunda.connector.generator.java.annotation.TemplateProperty;
-import io.camunda.connector.jdbc.model.auth.JdbcAuthentication;
-import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import java.util.List;
 
-public record JdbcRequest(
-    @NotNull
-        @TemplateProperty(
-            id = "database",
-            label = "Database",
-            group = "database",
-            type = Dropdown,
-            constraints = @TemplateProperty.PropertyConstraints(notEmpty = true),
-            choices = {
-              @TemplateProperty.DropdownPropertyChoice(label = "MSSQL", value = "MSSQL"),
-              @TemplateProperty.DropdownPropertyChoice(label = "MySQL", value = "MYSQL"),
-              @TemplateProperty.DropdownPropertyChoice(label = "PostgreSQL", value = "POSTGRESQL"),
-            })
-        SupportedDatabase database,
-    @Valid JdbcAuthentication authentication,
+public record JdbcRequestData(
     @NotNull
         @TemplateProperty(
             id = "isModifyingQuery",
