+++ connectors/jdbc/src/main/java/io/camunda/connector/jdbc/model/request/auth/DetailedConnection.java
@@ -4,16 +4,29 @@
  * See the License.txt file for more information. You may not use this file
  * except in compliance with the proprietary license.
  */
-package io.camunda.connector.jdbc.model.auth;
+package io.camunda.connector.jdbc.model.request.auth;
 
 import io.camunda.connector.generator.java.annotation.TemplateProperty;
 import io.camunda.connector.generator.java.annotation.TemplateSubType;
+import io.camunda.connector.jdbc.model.request.SupportedDatabase;
 import jakarta.validation.constraints.NotBlank;
+import java.util.Optional;
+import java.util.function.Function;
 
 @TemplateSubType(id = "detailed", label = "Detailed")
-public record DetailedAuthentication(
+public record DetailedConnection(
     @NotBlank @TemplateProperty(group = "authentication", label = "Host") String host,
     @NotBlank @TemplateProperty(group = "authentication", label = "Port") String port,
     @TemplateProperty(group = "authentication", label = "Username") String username,
     @TemplateProperty(group = "authentication", label = "Password") String password)
-    implements JdbcAuthentication {}
+    implements JdbcConnection {
+  @Override
+  public String getConnectionString(SupportedDatabase database) {
+    String authentication = Optional.ofNullable(username).map(buildUserPasswordString()).orElse("");
+    return database.getUrlSchema() + authentication + host + ":" + port;
+  }
+
+  private Function<String, String> buildUserPasswordString() {
+    return u -> u + ":" + password + "@";
+  }
+}
