+++ connectors/jdbc/src/main/java/io/camunda/connector/jdbc/model/request/auth/JdbcConnection.java
@@ -4,20 +4,23 @@
  * See the License.txt file for more information. You may not use this file
  * except in compliance with the proprietary license.
  */
-package io.camunda.connector.jdbc.model.auth;
+package io.camunda.connector.jdbc.model.request.auth;
 
 import com.fasterxml.jackson.annotation.JsonSubTypes;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
 import io.camunda.connector.generator.java.annotation.TemplateDiscriminatorProperty;
+import io.camunda.connector.jdbc.model.request.SupportedDatabase;
 
 @JsonTypeInfo(use = JsonTypeInfo.Id.NAME, property = "authType")
 @JsonSubTypes({
-  @JsonSubTypes.Type(value = UriAuthentication.class, name = "uri"),
-  @JsonSubTypes.Type(value = DetailedAuthentication.class, name = "detailed")
+  @JsonSubTypes.Type(value = UriConnection.class, name = "uri"),
+  @JsonSubTypes.Type(value = DetailedConnection.class, name = "detailed")
 })
 @TemplateDiscriminatorProperty(
     label = "Connection type",
     group = "authentication",
     name = "authType",
     defaultValue = "uri")
-public sealed interface JdbcAuthentication permits UriAuthentication, DetailedAuthentication {}
+public sealed interface JdbcConnection permits UriConnection, DetailedConnection {
+  String getConnectionString(SupportedDatabase database);
+}
