+++ connectors/jdbc/src/test/java/io/camunda/connector/jdbc/model/request/JdbcRequestTest.java
@@ -13,6 +13,7 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 import io.camunda.connector.api.error.ConnectorInputException;
 import io.camunda.connector.api.outbound.OutboundConnectorContext;
 import io.camunda.connector.jdbc.BaseTest;
+import io.camunda.connector.jdbc.model.request.JdbcRequest;
 import io.camunda.connector.validation.impl.DefaultValidationProvider;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
